12.0

+++ PrivateFrameworks/CompanionCamera.framework/CCCameraConnectionInternal.h
@@ -17,14 +17,14 @@
 - (oneway void)xpc_endBurstCaptureWithReply:(id /* block */)arg1;
 - (oneway void)xpc_ensureSwitchedToOneOfSupportedCaptureModes:(id)arg1 reply:(id /* block */)arg2;
 - (oneway void)xpc_fetchCurrentStateIncludingSupportedCaptureModes:(bool)arg1 reply:(id /* block */)arg2;
-- (oneway void)xpc_setCaptureMode:(int)arg1 reply:(id /* block */)arg2;
-- (oneway void)xpc_setFlashMode:(int)arg1;
+- (oneway void)xpc_setCaptureMode:(long long)arg1 reply:(id /* block */)arg2;
+- (oneway void)xpc_setFlashMode:(long long)arg1;
 - (oneway void)xpc_setFocusPoint:(id)arg1;
-- (oneway void)xpc_setHDRMode:(int)arg1;
-- (oneway void)xpc_setIrisMode:(int)arg1;
+- (oneway void)xpc_setHDRMode:(long long)arg1;
+- (oneway void)xpc_setIrisMode:(long long)arg1;
 - (oneway void)xpc_setPreviewEndpoint:(id)arg1;
 - (oneway void)xpc_setZoom:(float)arg1 reply:(id /* block */)arg2;
-- (oneway void)xpc_startCaptureWithMode:(int)arg1 reply:(id /* block */)arg2;
+- (oneway void)xpc_startCaptureWithMode:(long long)arg1 reply:(id /* block */)arg2;
 - (oneway void)xpc_stopCaptureWithReply:(id /* block */)arg1;
 - (oneway void)xpc_toggleCameraDevice;
