+++ PrivateFrameworks/MediaRemote.framework/MRGetVoiceInputDevicesResponseMessage.h
@@ -5,11 +5,11 @@
 @interface MRGetVoiceInputDevicesResponseMessage : MRProtocolMessage
 
 @property (nonatomic, readonly) NSArray *deviceIDs;
-@property (nonatomic, readonly) unsigned int errorCode;
+@property (nonatomic, readonly) long long errorCode;
 
 - (id)deviceIDs;
-- (unsigned int)errorCode;
-- (id)initWithDeviceIDs:(id)arg1 errorCode:(unsigned int)arg2;
+- (long long)errorCode;
+- (id)initWithDeviceIDs:(id)arg1 errorCode:(long long)arg2;
 - (unsigned long long)type;
 
 @end
