+++ PrivateFrameworks/SpotlightServices.framework/PRSL2FeatureVectorProcessingContext.h
@@ -4,17 +4,17 @@
 
 @interface PRSL2FeatureVectorProcessingContext : NSObject {
     unsigned long long  _expandedFeatureCount;
-    unsigned long long * _expandedFeatureExecutionOrder;
+    unsigned short * _expandedFeatureExecutionOrder;
 }
 
 @property (nonatomic) unsigned long long expandedFeatureCount;
-@property (nonatomic) unsigned long long*expandedFeatureExecutionOrder;
+@property (nonatomic) unsigned short*expandedFeatureExecutionOrder;
 
 - (void)dealloc;
 - (unsigned long long)expandedFeatureCount;
-- (unsigned long long*)expandedFeatureExecutionOrder;
+- (unsigned short*)expandedFeatureExecutionOrder;
 - (id)initWithFeatureOrder:(id)arg1 withInflation:(unsigned long long)arg2 withInflatedIndexToSize:(id)arg3;
 - (void)setExpandedFeatureCount:(unsigned long long)arg1;
-- (void)setExpandedFeatureExecutionOrder:(unsigned long long*)arg1;
+- (void)setExpandedFeatureExecutionOrder:(unsigned short*)arg1;
 
 @end
