+++ PrivateFrameworks/TimeSync.framework/TSgPTPMachPort.h
@@ -10,6 +10,6 @@
 - (void)dealloc;
 - (id)initWithMatchingDictionary:(id)arg1;
 - (id)initWithService:(unsigned int)arg1;
-- (long long)portType;
+- (int)portType;
 
 @end
