+++ PrivateFrameworks/TimeSync.framework/TSgPTPPort.h
@@ -9,14 +9,14 @@
     NSObject<OS_dispatch_queue> * _notificationsQueue;
     __TSgPTPPortNotification * _portForNotification;
     unsigned short  _portNumber;
-    long long  _portRole;
+    int  _portRole;
     NSObject<OS_dispatch_queue> * _propertyUpdateQueue;
     unsigned int  _service;
 }
 
 @property (nonatomic) unsigned short portNumber;
-@property (nonatomic) long long portRole;
-@property (nonatomic, readonly) long long portType;
+@property (nonatomic) int portRole;
+@property (nonatomic, readonly) int portType;
 @property (nonatomic, retain) NSObject<OS_dispatch_queue> *propertyUpdateQueue;
 @property (nonatomic, readonly) unsigned int service;
 
@@ -26,19 +26,19 @@
 
 - (bool)_commonInitWithService:(unsigned int)arg1;
 - (unsigned short)_portNumber;
-- (long long)_portRole;
+- (int)_portRole;
 - (void)dealloc;
 - (id)init;
 - (id)initWithMatchingDictionary:(id)arg1;
 - (id)initWithService:(unsigned int)arg1;
 - (unsigned short)portNumber;
-- (long long)portRole;
-- (long long)portType;
+- (int)portRole;
+- (int)portType;
 - (id)propertyUpdateQueue;
 - (unsigned int)service;
 - (void)serviceTerminated;
 - (void)setPortNumber:(unsigned short)arg1;
-- (void)setPortRole:(long long)arg1;
+- (void)setPortRole:(int)arg1;
 - (void)setPropertyUpdateQueue:(id)arg1;
 - (bool)startAutomaticPropertyUpdates;
 - (bool)stopAutomaticPropertyUpdates;
