+++ PrivateFrameworks/TimeSync.framework/TSgPTPUnicastLinkLayerEtEPort.h
@@ -13,6 +13,6 @@
 - (id)initWithInterfaceName:(id)arg1 andMACAddress:(char *)arg2;
 - (id)initWithMatchingDictionary:(id)arg1;
 - (id)initWithService:(unsigned int)arg1;
-- (long long)portType;
+- (int)portType;
 
 @end
