+++ PrivateFrameworks/iWorkImport.framework/TSTLayoutDynamicResizeInfoAccordion.h
@@ -16,7 +16,7 @@
 - (double)heightFactor;
 - (id)initWithDynamicResizeInfo:(id)arg1;
 - (id)initWithMasterLayout:(id)arg1 columnRegion:(id)arg2 rowRegion:(id)arg3;
-- (double)p_applyAccordionResizeFactor:(double)arg1 region:(id)arg2 direction:(long long)arg3 totalOfCaptured:(double)arg4 count:(unsigned short)arg5 capturedArray:(double*)arg6 currentArray:(double*)arg7 minimumArray:(double*)arg8;
+- (double)p_applyAccordionResizeFactor:(double)arg1 region:(id)arg2 direction:(long long)arg3 totalOfCaptured:(double)arg4 count:(unsigned int)arg5 capturedArray:(double*)arg6 currentArray:(double*)arg7 minimumArray:(double*)arg8;
 - (double)widthFactor;
 
 @end
