+++ protocols/PXCPLServiceUIDelegate.h
@@ -10,6 +10,6 @@
 @optional
 
 - (bool)serviceUI:(PXCPLServiceUI *)arg1 performAction:(long long)arg2;
-- (void)serviceUI:(PXCPLServiceUI *)arg1 progressDidChange:(double)arg2;
+- (void)serviceUI:(PXCPLServiceUI *)arg1 progressDidChange:(float)arg2;
 
 @end
